+++ src/applications/vaos/tests/new-appointment/components/ReviewPage/index.va-request.unit.spec.jsx
@@ -26,7 +26,7 @@ import {
   mockRequestSubmit,
   mockFacilityFetch,
 } from '../../../mocks/helpers';
-import { mockAppointmentSubmit } from '../../../mocks/helpers.vaos';
+import { mockAppointmentSubmit } from '../../../mocks/helpers.v2';
 import { getVAFacilityMock } from '../../../mocks/v0';
 
 const initialState = {
